requirements.txt
streamlit
pyngrok

torch
numpy
pandas

transformers>=4.41.0
accelerate>=0.31.0
bitsandbytes>=0.43.1
sentence-transformers

faiss-cpu

pymupdf
pdfplumber

pydantic>=2.0,<2.12.4
json-repair
tqdm
